Support wiping volume when they are deleted.
This prevents the recreation of a volume at the same place of a
preceding one, with access to the underlying volume metadata (partition
table, filesystem, etc.)

This should be backward compatible with the previous implementation: if
no wipe algorithm has been specified, the volume is deleted as before.
Otherwise, the volume is first wiped (this can take a while) then
deleted.